+++ markdown_to_mrkdwn/__init__.py
@@ -1,6 +1,7 @@
 # This file is required to make Python treat the directory as a package
 
 from .converter import SlackMarkdownConverter
+from .block_kit import BlockKitConverter
 
-__version__ = "0.1.5"
-__all__ = ["SlackMarkdownConverter"]
+__version__ = "0.2.0"
+__all__ = ["SlackMarkdownConverter", "BlockKitConverter"]